+++ tests/parser/test_parser.py
@@ -4,7 +4,6 @@ from typing import NamedTuple, Dict, Any, Sequence, Union, Tuple, Optional, Set,
 import pytest
 
 import typeit
-from typeit import parser as p
 from typeit import flags
 from typeit import schema
 from typeit.sums import Either
@@ -15,7 +14,7 @@ def test_type_with_unclarified_list():
         x: Sequence
         y: List
 
-    mk_main, serialize_main = p.type_constructor ^ X
+    mk_main, serialize_main = typeit.type_constructor ^ X
     x = mk_main({'x': [], 'y': []})
     x = mk_main({'x': [1], 'y': ['1']})
     assert x.x[0] == int(x.y[0])
@@ -30,8 +29,8 @@ def test_primitives_strictness():
         c: float
         d: bool
 
-    mk_x, serialize_x = p.type_constructor ^ X
-    mk_x_nonstrict, serialize_x_nonstrict = p.type_constructor & flags.NON_STRICT_PRIMITIVES ^ X
+    mk_x, serialize_x = typeit.type_constructor ^ X
+    mk_x_nonstrict, serialize_x_nonstrict = typeit.type_constructor & flags.NON_STRICT_PRIMITIVES ^ X
 
     data = {
         'a': '1',
@@ -71,7 +70,7 @@ def test_serialize_list():
     class X(NamedTuple):
         x: Union[None, Sequence[str]]
 
-    mk_x, serialize_x = p.type_constructor ^ X
+    mk_x, serialize_x = typeit.type_constructor ^ X
     data = {
         'x': ['str'],
     }
@@ -92,7 +91,7 @@ def test_serialize_union_lists():
     class X(NamedTuple):
         x: Union[Sequence[str], Sequence[float], Sequence[int]]
 
-    mk_x, serialize_x = p.type_constructor ^ X
+    mk_x, serialize_x = typeit.type_constructor ^ X
     data = {
         'x': [1],
     }
@@ -106,7 +105,7 @@ def test_type_with_sequence():
         y: Sequence[Any]
         z: Sequence[str]
 
-    mk_main, serializer = p.type_constructor(X)
+    mk_main, serializer = typeit.type_constructor(X)
 
     x = mk_main({'x': 1, 'y': [], 'z': ['Hello']})
     assert x.y == []
@@ -123,7 +122,7 @@ def test_type_with_tuple_primitives():
         b: Tuple            # the following are equivalent
         c: tuple
 
-    mk_x, serializer = p.type_constructor(X)
+    mk_x, serializer = typeit.type_constructor(X)
 
     x = mk_x({
         'a': ['value', 5],
@@ -167,7 +166,7 @@ def test_type_with_complex_tuples():
         a: Tuple[Tuple[Dict, Y], int]
         b: Optional[Any]
 
-    mk_x, serializer = p.type_constructor(X)
+    mk_x, serializer = typeit.type_constructor(X)
 
     x = mk_x({
         'a': [
@@ -194,7 +193,7 @@ def test_unsupported_variable_length_tuples():
         a: Tuple[int, ...]
 
     with pytest.raises(TypeError):
-        mk_x, serialize_x = p.type_constructor(X)
+        mk_x, serialize_x = typeit.type_constructor(X)
 
 
 def test_enum_like_types():
@@ -205,7 +204,7 @@ def test_enum_like_types():
     class X(NamedTuple):
         e: Enums
 
-    mk_x, serialize_x = p.type_constructor(X)
+    mk_x, serialize_x = typeit.type_constructor(X)
 
     data = {'e': 'a'}
     x = mk_x(data)
@@ -232,7 +231,7 @@ def test_sum_types_as_union():
     class X(NamedTuple):
         x: MyEither
 
-    mk_x, serialize_x = p.type_constructor ^ X
+    mk_x, serialize_x = typeit.type_constructor ^ X
     x_data = {
         'x': ('left', {'err': 'Error'})
     }
@@ -294,7 +293,7 @@ def test_enum_unions_serialization():
         val: Union[E0, E1]
 
 
-    __, serializer = p.type_constructor(MyType)
+    __, serializer = typeit.type_constructor(MyType)
 
     assert serializer(MyType(val=E1.Z)) == {'val': 'z'}
 
@@ -308,7 +307,7 @@ def test_type_with_empty_enum_variant():
         x: int
         y: Types
 
-    mk_x, serializer = p.type_constructor(X)
+    mk_x, serializer = typeit.type_constructor(X)
 
     for variant in Types:
         x = mk_x({'x': 1, 'y': variant.value})
@@ -329,7 +328,7 @@ def test_type_with_set():
         g: Set[Any]
         h: Set[int]
 
-    mk_x, serializer = p.type_constructor(X)
+    mk_x, serializer = typeit.type_constructor(X)
 
     x = mk_x({
         'a': [],
@@ -356,7 +355,7 @@ def test_parse_sequence():
     XS = Sequence[X]
 
     data = [{'x': 1, 'y': {}}]
-    mk_xs, serialize_xs = p.type_constructor(XS)
+    mk_xs, serialize_xs = typeit.type_constructor(XS)
     z = mk_xs(data)
     assert z[0].x == 1
     assert serialize_xs(z) == data
@@ -365,7 +364,7 @@ def test_parse_sequence():
     XS = Sequence[int]
 
     data = [1, 2, 3]
-    mk_xs, serialize_xs = p.type_constructor(XS)
+    mk_xs, serialize_xs = typeit.type_constructor(XS)
 
     z = mk_xs(data)
     assert z[0] == 1
@@ -379,7 +378,7 @@ def test_parse_sequence():
     (Dict[str, Any], {'x': 1, 'y': True, 'z': '1'})
 ))
 def test_parse_builtins(typ, data):
-    mk_x, serialize_x = p.type_constructor(typ)
+    mk_x, serialize_x = typeit.type_constructor(typ)
 
     z = mk_x(data)
     assert z == data
@@ -399,7 +398,7 @@ def test_type_with_dict():
         x: int
         y: Dict[str, Any]
 
-    mk_x, serializer = p.type_constructor(X)
+    mk_x, serializer = typeit.type_constructor(X)
 
     with pytest.raises(typeit.Error):
         mk_x({})
@@ -418,10 +417,10 @@ def test_name_overrides():
     data = {'my-x': 1}
 
     with pytest.raises(typeit.Error):
-        mk_x, serialize_x = p.type_constructor ^ X
+        mk_x, serialize_x = typeit.type_constructor ^ X
         mk_x(data)
 
-    mk_x, serialize_x = p.type_constructor & {X.x: 'my-x'} ^ X
+    mk_x, serialize_x = typeit.type_constructor & {X.x: 'my-x'} ^ X
     x = mk_x(data)
     assert serialize_x(x) == data
 
@@ -435,4 +434,4 @@ def test_name_overrides():
     Sequence,
 ])
 def test_parse_builtins_and_sequences(typ):
-    mk_x, serialize_x = p.type_constructor ^ typ
+    mk_x, serialize_x = typeit.type_constructor ^ typ