+++ typeit/schema/__init__.py
@@ -2,5 +2,6 @@ from . import meta
 from . import primitives
 from . import types
 from . import nodes
+from .errors import Invalid
 
-__all__ = ['meta', 'primitives', 'types', 'nodes']
+__all__ = ['meta', 'primitives', 'types', 'nodes', 'Invalid']